(re)move old (compile-time) platform code
